refactor: Spring 3.x upgrade (wip)

+++ redis-om-spring/src/main/java/com/redis/om/spring/metamodel/nonindexed/NonIndexedTagField.java
@@ -22,12 +22,12 @@ public class NonIndexedTagField<E, T> extends MetamodelField<E, T> {
     return new ArrayAppendAction<>(field, value);
   }
 
-  public Consumer<? super E> insert(Object value, Long index) {
+  public Consumer<? super E> insert(Object value, Integer index) {
     return new ArrayInsertAction<>(field, value, index);
   }
 
   public Consumer<? super E> prepend(Object value) {
-    return new ArrayInsertAction<>(field, value, 0L);
+    return new ArrayInsertAction<>(field, value, 0);
   }
 
   public ToLongFunction<? super E> length() {
@@ -38,27 +38,27 @@ public class NonIndexedTagField<E, T> extends MetamodelField<E, T> {
     return new ArrayIndexOfAction<>(field, element);
   }
 
-  public <R> ArrayPopAction<? super E,R> pop(Long index) {
+  public <R> ArrayPopAction<? super E,R> pop(Integer index) {
     return new ArrayPopAction<>(field, index);
   }
 
   public <R> ArrayPopAction<? super E,R> pop() {
-    return pop(-1L);
+    return pop(-1);
   }
 
   public <R> ArrayPopAction<? super E,R> removeFirst() {
-    return pop(0L);
+    return pop(0);
   }
 
   public <R> ArrayPopAction<? super E,R> removeLast() {
-    return pop(-1L);
+    return pop(-1);
   }
 
-  public <R> ArrayPopAction<? super E,R> remove(Long index) {
+  public <R> ArrayPopAction<? super E,R> remove(Integer index) {
     return pop(index);
   }
 
-  public Consumer<? super E> trimToRange(Long begin, Long end) {
+  public Consumer<? super E> trimToRange(Integer begin, Integer end) {
     return new ArrayTrimAction<>(field, begin, end);
   }
